+++ src/types/CartItems.ts
@@ -1,0 +1,1 @@
+export type CartItems = { [itemID: number]: { quantity: number } };
