--- src/types/CartList.ts
@@ -1,1 +1,0 @@
-export type CartList = { [key: number]: { quantity: number } };
